+++ src/client.ts
@@ -1,16 +1,16 @@
-import { ProjectsEndpoint } from "./endpoints/projects";
-import { UsersEndpoint } from "./endpoints/users";
+import { CompanyEndpoint } from "./endpoints/company";
+import { LookupEndpoint } from "./endpoints/lookup";
 import { HttpClient } from "./http";
 import type { ApiClientConfig } from "./types";
 
 export class ApiClient {
-  public readonly users: UsersEndpoint;
-  public readonly projects: ProjectsEndpoint;
+  public readonly company: CompanyEndpoint;
+  public readonly lookup: LookupEndpoint;
   private readonly http: HttpClient;
 
   constructor(config: ApiClientConfig) {
     this.http = new HttpClient(config);
-    this.users = new UsersEndpoint(this.http);
-    this.projects = new ProjectsEndpoint(this.http);
+    this.company = new CompanyEndpoint(this.http);
+    this.lookup = new LookupEndpoint(this.http);
   }
 }